requirements.txt
mindspore==2.1.0
pylint==2.17.4


numpy~=1.24.3
scipy~=1.10.1
argparse~=1.4.0
matplotlib~=3.7.1

pillow~=9.4.0
mindcv~=0.2.2

